Expanded Hemophage Food Options
Adds the "BLOODY" flag to a bunch of pre-existing food so that hemophages can eat it.  Food selected is based on it either containing blood, dishes where at least half the ingredients are organs, human meat (for the cannibal gimmicks out there), and yet another media reference.

Currently drafted because there's a hell of a lot of food I want to go through, as well as take suggestions for what to add/remove.

Plan to also add a loadout pill bottle of taste suppressor, since it's a niche chem that I'd imagine hemophages trying to blend in or hide the affliction would be familiar with.